+++ packages/javi-client/index.ts
@@ -1,3 +1,4 @@
+export * from "./DevTemplate";
 export * from "./getClientConf";
 export * from "./JaviDirector";
 export * from "./types";