+++ packages/misc/node/index.ts
@@ -1,4 +1,5 @@
 export * from "./createWebpackBaseConf";
 export * from "./test-enzyme";
 export * from "./test-react-test-renderer";
+export * from "./ts-node-util";
 export * from "./webpack";