+++ packages/javi/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## :tada: Enhancements
 
+- Trimmed dependencies
+
 ## :bug: Bug fixes
 
 ## :boom: Breaking changes